RAN-023-connection-based-on-coverage-radius
- implemented the connection of each UE to
  the GNB based on the GNB coverage radius
- implemented this functionality using RadioHub
  methods, as a simulator for transmitting
  radio signals over the air
- shortened the name UeRrcState::SEARCHING_FOR_CELL
  for better display on the screen
- implemented handleDeregistration method
  on RadioHub side

+++ gnb/src/gnb_logic.cpp
@@ -5,7 +5,7 @@
 
 #include "flow_logger.hpp"
 
-GnbLogic::GnbLogic(uint32_t id, uint16_t radius, QObject* parent)
+GnbLogic::GnbLogic(uint32_t id, double radius, QObject* parent)
     : BaseEntity(id, EntityType::GNB, parent)
     , radius_(radius)
 {
@@ -33,7 +33,7 @@ uint32_t GnbLogic::getConnectedUeCount() const
     return static_cast<uint32_t>(ue_contexts_.size());
 }
 
-uint16_t GnbLogic::getRadius() const
+double GnbLogic::getRadius() const
 {
     return radius_;
 }
